Allow for config driven service locator
This is not ideal, but Spring doesn't seem to have a nice way for the
`SpringApplication` to override an implementation on start up that I can
find. Simple access to the IoC container would allow for a must nicer
solution by simply registering a different implementation for the
application on start up but can't find anywhere in spring to do this.