Built new indexing tool due to bug, updated scripts
The changes made to the indexing tool are just switching a dependency
that a) Shared package/class name and b) different version were imported
in the parent vs a child pom. Everything would build fine but at runtime
different versions of a type was being loaded than expected due to this
collision. Update I made can be found at
http://github.com/Layoric/stanbol/tree/0.12-index-tool-fix.